Make example tests reusable across tree implementations.
